Execute clean commands in a subshell
Each clean command alters the working directory with `cd` and does not reset it. This means that if several project directories are found (identified by relative paths), the first directory will be cleaned, but all subsequent directories will fail now that the project paths are no longer valid relative to the altered working directory.

Solve this by wrapping the `eval` clean command statements in parentheses, [which will cause them to be executed in a subshell](http://superuser.com/a/231884). This means the working directory will not be altered in the master loop.